Part 2 This is a major update to the Project files.
I created a new project called Script Library which is actually Secrets of the Stones. I temp named it this for the reason as it is a complete rewrite of all scripts.

As of right now it is broken in many areas.

I am in the process of auditing the scripts as well as meshes/textures.

I have found many mistakes in the meshes and fixed the paths with NifSkope.

The Mesh Pack section is for New Weapons and New Armor that I will be add into the Secrets of the Stones project as soon as I finish Auditing all the textures and meshes.

I am making it a separate esp called Mesh Pack.esp which I am creating the items in game using Ebony and Glass armor as the bases for the stats. once I get it completed and Audited I will be merging it into the main esp.

There has been a update to the Script section of the SVN adding in new scripts to the projects.

Once I get the Audit done on the content that is already here I will move on to adding in more of what is on the to do list.

A lot of the meshes and used in this project have been made by many people and if your name is not listed please inform me so that I can add you to this list.
Most of the meshes in this mesh pack are over 3 years old and the Arthur information has been lost in the sands of time.

git-svn-id: https://svn.code.sf.net/p/oblivionworks/code@4 c80116c5-1073-4541-b2a9-63b9fd661f84