A Complete Audit of all of the Meshes In Game have been done.
The AUDIT File contains all of the Objects in game that is added.

I will begin a Audit of all the Scripts soon

I started on the restructure of the Arcane Bank making changes to how things are handled and will have that section completed by the end of the week.

git-svn-id: https://svn.code.sf.net/p/oblivionworks/code@8 c80116c5-1073-4541-b2a9-63b9fd661f84