Added Support to tes4files
Moved the Default Locations of TES4Gecko to the TES4 Gecko folder in the Bethesda Softworks folder to help keep the oblivion Folder clean and so that it will not be scanned by the oblivion,exe during game play.

Moved the TES4Edit as ell to the TES4Edit folder which will be located again in the Bethesda Softworks folder to help keep the oblivion Folder clean and so that it will not be scanned by the oblivion,exe during game play.

updated the help html file to reflect the changes as well as another minor change to teh ini file to explain the change.

git-svn-id: https://svn.code.sf.net/p/oblivionworks/code@55 c80116c5-1073-4541-b2a9-63b9fd661f84