Some personal changes that have really made programming/debugging easier for me, but not really appropriate for general users.  So I've hidden them with an ini setting 'bShowDevTools'
The first is a 'Restart' status bar button.  Very useful for testing something out, editing the code, then restarting to check out if the changes worked or not.

The second is a mod menu command 'Test Full Record Definitions...', which simply loads the selected mod using Python mode, forcing the use of every available record reader.  Useful to use on Skyrim.esm to verify all definitions are correct and don't cause errors.  Also useful when a user reports a problem with a mod / a bug report involving a mod, run the command on that mod instead of trying to rebuild a Bashed Patch - much quicker.

git-svn-id: https://svn.code.sf.net/p/oblivionworks/code@2273 c80116c5-1073-4541-b2a9-63b9fd661f84